+++ statickg/models/prelude.py
@@ -1,4 +1,4 @@
-from statickg.models.etl import Change, ETLConfig, ETLFileTracker, ETLTask, Service
+from statickg.models.etl import Change, ETLConfig, ETLOutput, ETLTask, Service
 from statickg.models.input_file import (
     BaseType,
     InputFile,
@@ -12,7 +12,7 @@ from statickg.models.repository import GitRepository, Repository
 __all__ = [
     "ETLConfig",
     "ETLTask",
-    "ETLFileTracker",
+    "ETLOutput",
     "Change",
     "Service",
     "InputFile",